requirements.txt
Flask
gunicorn
python-dotenv
google-generativeai
googletrans==4.0.0-rc1
requests
psycopg2-binary